Modify null clock handling in picker
The plugin crash protection layer is predicated on ClockRegistry
returning null when a plugin has a crashing problem. As a result,
ThemePicker requiring a non-null response appears to hide the
underlying error. This fixes that by handling the null value at
all the relevant callsites by doing nothing.

The clock parts of ThemePicker will still undoubtedly be broken in
some way since a clock is failing to load or function, but this is
preferrable to ThemePicker crashing as a result of changes to the
library, especially since picker presubmits are not currently
running there.

Bug: 398070481
Test: Manual + Presubmits
Flag: NONE Bugfix
Change-Id: Id8e6f9888bc3b84c084587b2b4adade3934de286